+++ src/components/shared/Game/PlayField/CancelActionButton.vue
@@ -44,6 +44,11 @@ export default {
         noTargetSelectedText() {
             if (this.game.isFirstWaitingChooseSideAction) {
                 return this.$t("CancelActionButton.pleaseChooseSide");
+            } else if (this.game.isFirstWaitingChooseCardAction) {
+                if (this.game.isFirstWaitingSkippableAction) {
+                    return this.$t("CancelActionButton.chooseTCardIfNot");
+                }
+                return this.$t("CancelActionButton.pleaseChooseCard");
             } else if (this.game.isFirstWaitingSkippableAction) {
                 return this.$t("CancelActionButton.chooseTargetIfNot");
             }